+++ SO_2SEM/Semaforo.h
@@ -26,7 +26,7 @@ void *consumer(void *cno);
  * \param placar Ponteiro para o placar.
  * \param pontuacao Ponteiro para a pontuação a ser adicionada.
 */
-void enter_item(float* placar, int* pontuacao);
+void enter_item(int* placar, int* pontuacao);
 
 /**
  * \brief Consome um item e armazena em placar utilizando a região crítica.
@@ -34,7 +34,7 @@ void enter_item(float* placar, int* pontuacao);
  * \param placar Ponteiro para o placar.
  * \param pontuacao Ponteiro para a pontuação a ser removida.
 */
-void remove_item(float* placar, int* pontuacao);
+void remove_item(int* placar, int* pontuacao);
 
 /**
  * \brief Inicia o semafóro empty com N, o semafóro full com 0, e o mutex como desbloqueado.
